+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## 0.30.0 (PENDING)
 
 - Add `<x-titlebar>` element
+- Add `<x-pager>` element
 
 ## 0.29.0 (2025-02-02)